--- hazelcast-jet-sql/src/main/java/com/hazelcast/jet/sql/impl/aggregate/IgnoreNullsSqlAggregation.java
@@ -1,68 +1,0 @@
-/*
- * Copyright (c) 2008-2020, Hazelcast, Inc. All Rights Reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.hazelcast.jet.sql.impl.aggregate;
-
-import com.hazelcast.nio.ObjectDataInput;
-import com.hazelcast.nio.ObjectDataOutput;
-
-import javax.annotation.concurrent.NotThreadSafe;
-import java.io.IOException;
-
-@NotThreadSafe
-class IgnoreNullsSqlAggregation implements SqlAggregation {
-
-    private SqlAggregation delegate;
-
-    @SuppressWarnings("unused")
-    private IgnoreNullsSqlAggregation() {
-    }
-
-    IgnoreNullsSqlAggregation(SqlAggregation delegate) {
-        this.delegate = delegate;
-    }
-
-    @Override
-    public void accumulate(Object value) {
-        if (value == null) {
-            return;
-        }
-
-        delegate.accumulate(value);
-    }
-
-    @Override
-    public void combine(SqlAggregation other0) {
-        IgnoreNullsSqlAggregation other = (IgnoreNullsSqlAggregation) other0;
-
-        delegate.combine(other.delegate);
-    }
-
-    @Override
-    public Object collect() {
-        return delegate.collect();
-    }
-
-    @Override
-    public void writeData(ObjectDataOutput out) throws IOException {
-        out.writeObject(delegate);
-    }
-
-    @Override
-    public void readData(ObjectDataInput in) throws IOException {
-        delegate = in.readObject();
-    }
-}
